Many changes, spurred by the need to submit #1:
1.  Wrap EVAL-WHEN around the optimization declaration at the head of
the file.  This keeps the optimization settings from "leaking out" of
the compilation of SHOP2.

2.  Added :plans as a SHOP-TRACE option.

3.  Modified FIND-SATISFIERS to pass through a LEVEL argument.  The
level was incorrectly being reset to 0 when this was called
recursively.

4.  Added *break-on-backtrack* global.

5.  Modified operator tracing to print more information about the
state before and after the operator application.

6.  Allow the user to keep MAKE-PROBLEM from printing stuff to the
screen.

7.  Fix bugs in plan saving and add :plans tracing.

8.  Check for missing method definitions and raise an error
condition.  This *could* be problematic if a user were expecting to be
able to cons together method definitions or something.  But it seemed
more useful to permit the user to get an error if s/he misspelled a
method name, used an as-yet-unprogrammed or deleted method name.

9.  Check syntax of FORALL postconditions.  These are actually
somewhat irregular vis a vis the same construct in preconditions, so I
added a check to raise an error on ill-formed ones.