Summary: Fixed logic for choosing domain.
The logic for choosing the domain in the current version of the function was to
use the DOMAIN argument first, then get the domain from the PROBLEM.

But the DOMAIN defaults to *DOMAIN*, which is set by DEFDOMAIN, so the DOMAIN
argument is ALWAYS bound, unless there are no domains at all.  This doesn't make
much sense.

So I modified this to add a DOMAIN-SUPPLIED-P argument, and now the logic is:
1. If the domain argument is explicitly passed, use that;
2. If the domain is specified in the PROBLEM, use that; 
3. Use the value of *DOMAIN*.

I think the current approach could be simplified and cleaned up, but this is
much better than what was there, so am committing this now.


git-svn-id: https://svn.sift.info:3333/svn/shop2-public/shop/trunk/shop2@238 42e745e6-6fc8-43ef-911e-54bb78f6b639