Merge the EXPLICIT-STATE-SEARCH branch.
Now we have an alternative way of doing search for SHOP2 that uses an
explicit stack, rather than using the program stack as the search
stack as in standard SHOP2.

This merge should not in any way interfere with the normal functioning
of SHOP2, since it provides a separate entry point.

We are not certain that backtracking works correctly in this version
of SHOP2, so the explicit-state-search version should be viewed as
only a proof-of-concept for now.


git-svn-id: https://svn.sift.info:3333/svn/shop2-public/shop/trunk/shop2@305 42e745e6-6fc8-43ef-911e-54bb78f6b639