Fix problem with finding :INCLUDE files created by ASDF.
ASDF moves FASL files into a cache directory, which means that *load-truename*
doesn't point to a file in the directory with the source code, but instead
points to a file in the cache directory.  That was breaking the use of relative
pathnames in :INCLUDE in DEFDOMAIN.

Fixed this by adding to the DEFDOMAIN a cached copy of the filename that will be
cached at compile time (at macro-expansion time), instead of at load time when
that redirection has occurred.


git-svn-id: https://svn.sift.info:3333/svn/shop2-public/shop/trunk/shop2@331 42e745e6-6fc8-43ef-911e-54bb78f6b639